add authenticate middleware

+++ noteTakingApp/src/controllers/notes/destroy.php
@@ -10,13 +10,18 @@ $heading = "My Notes";
 
 $id = explode("/", $_SERVER['REQUEST_URI'])[2];;
 
-if (! isset($note)) {
-    Router::abort(Response::FORBIDDEN);
-}
-authorize($note['userId'] === $_SESSION['user'], Response::FORBIDDEN);
-
-$note = $db->delete("DELETE FROM notes WHERE id = :id", [
+$note = $db->query("SELECT * FROM notes WHERE id = :id", [
     'id' => $id,
+])->findOrFail();
+
+if (! isset($note)) {
+    Router::abort(Response::NOT_FOUND);
+}
+authorize($note['userId'] == $_SESSION['user'], Response::FORBIDDEN);
+
+$note = $db->delete("DELETE FROM notes WHERE id = :id AND userId = :user", [
+    'id' => $id,
+    'user' => $_SESSION['user'],
 ]);
 
 header("Location: /notes");
